+++ src/test/java/com/cosmian/TestFfiFindex.java
@@ -6,31 +6,22 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import java.nio.ByteBuffer;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
-import java.sql.SQLException;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
-import java.util.stream.Collectors;
 
 import org.apache.commons.lang3.ArrayUtils;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
+import com.cosmian.findex.Redis;
 import com.cosmian.findex.Sqlite;
 import com.cosmian.findex.UsersDataset;
-import com.cosmian.jna.FfiException;
 import com.cosmian.jna.findex.Ffi;
 import com.cosmian.jna.findex.IndexedValue;
 import com.cosmian.jna.findex.Location;
 import com.cosmian.jna.findex.MasterKeys;
 import com.cosmian.jna.findex.Word;
-import com.cosmian.jna.findex.Callbacks.FetchAllEntry;
-import com.cosmian.jna.findex.Callbacks.FetchChain;
-import com.cosmian.jna.findex.Callbacks.FetchEntry;
-import com.cosmian.jna.findex.Callbacks.ListRemovedLocations;
-import com.cosmian.jna.findex.Callbacks.UpdateLines;
-import com.cosmian.jna.findex.Callbacks.UpsertChain;
-import com.cosmian.jna.findex.Callbacks.UpsertEntry;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 public class TestFfiFindex {
@@ -47,10 +38,10 @@ public class TestFfiFindex {
     }
 
     @Test
-    public void testUpsertAndSearch() throws Exception {
+    public void testUpsertAndSearchSqlite() throws Exception {
         System.out.println("");
         System.out.println("---------------------------------------");
-        System.out.println("Findex Upsert");
+        System.out.println("Findex Upsert Sqlite");
         System.out.println("---------------------------------------");
         System.out.println("");
 
@@ -91,98 +82,9 @@ public class TestFfiFindex {
         db.insertUsers(testFindexDataset);
 
         //
-        // Declare all callbacks
-        //
-        FetchEntry fetchEntry = new FetchEntry(new com.cosmian.jna.findex.FfiWrapper.FetchEntryInterface() {
-            @Override
-            public HashMap<byte[], byte[]> fetch(List<byte[]> uids) throws FfiException {
-                try {
-                    return db.fetchEntryTableItems(uids);
-                } catch (SQLException e) {
-                    throw new FfiException("Failed fetch entry: " + e.toString());
-                }
-            }
-        });
-
-        FetchAllEntry fetchAllEntry = new FetchAllEntry(new com.cosmian.jna.findex.FfiWrapper.FetchAllEntryInterface() {
-            @Override
-            public HashMap<byte[], byte[]> fetch() throws FfiException {
-                try {
-                    return db.fetchAllEntryTableItems();
-                } catch (SQLException e) {
-                    throw new FfiException("Failed fetch all entry: " + e.toString());
-                }
-            }
-        });
-
-        FetchChain fetchChain = new FetchChain(new com.cosmian.jna.findex.FfiWrapper.FetchChainInterface() {
-            @Override
-            public HashMap<byte[], byte[]> fetch(List<byte[]> uids) throws FfiException {
-                try {
-                    return db.fetchChainTableItems(uids);
-                } catch (SQLException e) {
-                    throw new FfiException("Failed fetch chain: " + e.toString());
-                }
-            }
-        });
-
-        UpsertEntry upsertEntry = new UpsertEntry(new com.cosmian.jna.findex.FfiWrapper.UpsertEntryInterface() {
-            @Override
-            public void upsert(HashMap<byte[], byte[]> uidsAndValues) throws FfiException {
-                try {
-                    db.databaseUpsert(uidsAndValues, "entry_table");
-                } catch (SQLException e) {
-                    throw new FfiException("Failed entry upsert: " + e.toString());
-                }
-            }
-        });
-        UpsertChain upsertChain = new UpsertChain(new com.cosmian.jna.findex.FfiWrapper.UpsertChainInterface() {
-            @Override
-            public void upsert(HashMap<byte[], byte[]> uidsAndValues) throws FfiException {
-                try {
-                    db.databaseUpsert(uidsAndValues, "chain_table");
-                } catch (SQLException e) {
-                    throw new FfiException("Failed chain upsert: " + e.toString());
-                }
-            }
-        });
-        UpdateLines updateLines = new UpdateLines(new com.cosmian.jna.findex.FfiWrapper.UpdateLinesInterface() {
-            @Override
-            public void update(List<byte[]> removedChains, HashMap<byte[], byte[]> newEntries,
-                HashMap<byte[], byte[]> newChains) throws FfiException {
-                try {
-                    db.databaseTruncate("entry_table");
-                    db.databaseUpsert(newEntries, "entry_table");
-                    db.databaseUpsert(newChains, "chain_table");
-                    db.databaseRemove(removedChains, "chain_table");
-                } catch (SQLException e) {
-                    throw new FfiException("Failed update lines: " + e.toString());
-                }
-            }
-        });
-        ListRemovedLocations listRemovedLocations =
-            new ListRemovedLocations(new com.cosmian.jna.findex.FfiWrapper.ListRemovedLocationsInterface() {
-                @Override
-                public List<Location> list(List<Location> locations) throws FfiException {
-                    List<Integer> ids =
-                        locations.stream().map((Location location) -> ByteBuffer.wrap(location.getBytes()).getInt())
-                            .collect(Collectors.toList());
-
-                    try {
-                        return db.listRemovedIds("users", ids).stream()
-                            .map((Integer id) -> new Location(ByteBuffer.allocate(32).putInt(id).array()))
-                            .collect(Collectors.toList());
-                    } catch (SQLException e) {
-                        throw new FfiException("Failed update lines: " + e.toString());
-                    }
-
-                }
-            });
-
-        //
         // Upsert
         //
-        Ffi.upsert(masterKeys, label, indexedValuesAndWords, fetchEntry, upsertEntry, upsertChain);
+        Ffi.upsert(masterKeys, label, indexedValuesAndWords, db.fetchEntry, db.upsertEntry, db.upsertChain);
         System.out.println("After insertion: entry_table: nb indexes: " + db.getAllKeyValueItems("entry_table").size());
         System.out.println("After insertion: chain_table: nb indexes: " + db.getAllKeyValueItems("chain_table").size());
 
@@ -191,27 +93,26 @@ public class TestFfiFindex {
         //
         System.out.println("");
         System.out.println("---------------------------------------");
-        System.out.println("Findex Search");
+        System.out.println("Findex Search Sqlite");
         System.out.println("---------------------------------------");
         System.out.println("");
 
         {
             List<byte[]> indexedValuesList =
-                Ffi.search(masterKeys.getK(), label, new Word[] {new Word("France")}, 0, fetchEntry, fetchChain);
+                Ffi.search(masterKeys.getK(), label, new Word[] {new Word("France")}, 0, db.fetchEntry, db.fetchChain);
             int[] dbUids = indexedValuesBytesListToArray(indexedValuesList);
 
             assertArrayEquals(expectedDbUids, dbUids);
         }
 
         // This compact should do nothing except changing the label since the users table didn't change.
-        Ffi.compact(1, masterKeys, "NewLabel".getBytes(), fetchEntry, fetchChain, fetchAllEntry, updateLines,
-            listRemovedLocations);
+        Ffi.compact(1, masterKeys, "NewLabel".getBytes(), db.fetchEntry, db.fetchChain, db.fetchAllEntry,
+            db.updateLines, db.listRemovedLocations);
 
         {
             // Search with old label
-
             List<byte[]> indexedValuesList =
-                Ffi.search(masterKeys.getK(), label, new Word[] {new Word("France")}, 0, fetchEntry, fetchChain);
+                Ffi.search(masterKeys.getK(), label, new Word[] {new Word("France")}, 0, db.fetchEntry, db.fetchChain);
             int[] dbUids = indexedValuesBytesListToArray(indexedValuesList);
 
             assertEquals(0, dbUids.length);
@@ -219,9 +120,8 @@ public class TestFfiFindex {
 
         {
             // Search with new label and without user changes
-
             List<byte[]> indexedValuesList = Ffi.search(masterKeys.getK(), "NewLabel".getBytes(),
-                new Word[] {new Word("France")}, 0, fetchEntry, fetchChain);
+                new Word[] {new Word("France")}, 0, db.fetchEntry, db.fetchChain);
             int[] dbUids = indexedValuesBytesListToArray(indexedValuesList);
 
             assertArrayEquals(expectedDbUids, dbUids);
@@ -231,18 +131,150 @@ public class TestFfiFindex {
         db.deleteUser(17);
         int[] newExpectedDbUids = ArrayUtils.removeElement(expectedDbUids, 17);
 
-        Ffi.compact(1, masterKeys, "NewLabel".getBytes(), fetchEntry, fetchChain, fetchAllEntry, updateLines,
-            listRemovedLocations);
+        Ffi.compact(1, masterKeys, "NewLabel".getBytes(), db.fetchEntry, db.fetchChain, db.fetchAllEntry,
+            db.updateLines, db.listRemovedLocations);
 
         {
             // Search should return everyone instead of n°17
 
             List<byte[]> indexedValuesList = Ffi.search(masterKeys.getK(), "NewLabel".getBytes(),
-                new Word[] {new Word("France")}, 0, fetchEntry, fetchChain);
+                new Word[] {new Word("France")}, 0, db.fetchEntry, db.fetchChain);
             int[] dbUids = indexedValuesBytesListToArray(indexedValuesList);
 
             assertArrayEquals(newExpectedDbUids, dbUids);
         }
+
+        // Check allocation problem during insertions. Allocation problem could occur when fetching entry table
+        // values
+        // whose sizes depend on words being indexed: the Entry Table Encrypted value is:
+        // `EncSym(𝐾value, (ict_uid𝑥𝑤𝑖, 𝐾𝑤𝑖 , 𝑤𝑖))`
+        for (int i = 0; i < 100; i++) {
+            Ffi.upsert(masterKeys, label, indexedValuesAndWords, db.fetchEntry, db.upsertEntry, db.upsertChain);
+        }
+    }
+
+    @Test
+    public void testUpsertAndSearchRedis() throws Exception {
+        System.out.println("");
+        System.out.println("---------------------------------------");
+        System.out.println("Findex Upsert Redis");
+        System.out.println("---------------------------------------");
+        System.out.println("");
+
+        //
+        // Recover master keys (k and k*)
+        //
+        String masterKeysJson = Resources.load_resource("findex/keys.json");
+        MasterKeys masterKeys = MasterKeys.fromJson(masterKeysJson);
+
+        byte[] label = Resources.load_resource_as_bytes("findex/label");
+
+        //
+        // Recover test vectors
+        //
+        ObjectMapper mapper = new ObjectMapper();
+        String expectedSearchResultsInt = Resources.load_resource("findex/expected_db_uids.json");
+        int[] expectedDbUids = mapper.readValue(expectedSearchResultsInt, int[].class);
+        Arrays.sort(expectedDbUids);
+
+        //
+        // Build dataset with DB uids and words
+        //
+        String dataJson = Resources.load_resource("findex/data.json");
+        UsersDataset[] testFindexDataset = UsersDataset.fromJson(dataJson);
+        HashMap<IndexedValue, Word[]> indexedValuesAndWords = new HashMap<>();
+        for (UsersDataset user : testFindexDataset) {
+            ByteBuffer dbuf = ByteBuffer.allocate(32);
+            dbuf.putInt(user.id);
+            byte[] dbUid = dbuf.array();
+
+            indexedValuesAndWords.put(new Location(dbUid), user.values());
+        }
+
+        //
+        // Prepare Sqlite tables and users
+        //
+        // Sqlite db = new Sqlite();
+        // db.insertUsers(testFindexDataset);
+        Redis db = new Redis();
+        db.jedis.flushAll();
+        db.insertUsers(testFindexDataset);
+
+        //
+        // Upsert
+        //
+        Ffi.upsert(masterKeys, label, indexedValuesAndWords, db.fetchEntry, db.upsertEntry, db.upsertChain);
+        System.out.println("After insertion: entry_table: nb indexes: "
+            + db.getAllKeysAndValues(Redis.INDEX_TABLE_ENTRY_STORAGE).size());
+        System.out.println("After insertion: chain_table: nb indexes: "
+            + db.getAllKeysAndValues(Redis.INDEX_TABLE_CHAIN_STORAGE).size());
+
+        //
+        // Search
+        //
+        System.out.println("");
+        System.out.println("---------------------------------------");
+        System.out.println("Findex Search Redis");
+        System.out.println("---------------------------------------");
+        System.out.println("");
+
+        {
+            List<byte[]> indexedValuesList =
+                Ffi.search(masterKeys.getK(), label, new Word[] {new Word("France")}, 0, db.fetchEntry, db.fetchChain);
+            int[] dbUids = indexedValuesBytesListToArray(indexedValuesList);
+
+            assertArrayEquals(expectedDbUids, dbUids);
+        }
+
+        // This compact should do nothing except changing the label since the users table didn't change.
+        Ffi.compact(1, masterKeys, "NewLabel".getBytes(), db.fetchEntry, db.fetchChain, db.fetchAllEntry,
+        db.updateLines, db.listRemovedLocations);
+
+        {
+        // Search with old label
+
+        List<byte[]> indexedValuesList =
+        Ffi.search(masterKeys.getK(), label, new Word[] {new Word("France")}, 0, db.fetchEntry, db.fetchChain);
+        int[] dbUids = indexedValuesBytesListToArray(indexedValuesList);
+
+        assertEquals(0, dbUids.length);
+        }
+
+        {
+        // Search with new label and without user changes
+
+        List<byte[]> indexedValuesList = Ffi.search(masterKeys.getK(), "NewLabel".getBytes(),
+        new Word[] {new Word("France")}, 0, db.fetchEntry, db.fetchChain);
+        int[] dbUids = indexedValuesBytesListToArray(indexedValuesList);
+
+        assertArrayEquals(expectedDbUids, dbUids);
+        }
+
+        // Delete the user n°17 to test the compact indexes
+        db.deleteUser(17);
+        int[] newExpectedDbUids = ArrayUtils.removeElement(expectedDbUids, 17);
+
+        Ffi.compact(1, masterKeys, "NewLabel".getBytes(), db.fetchEntry, db.fetchChain, db.fetchAllEntry,
+        db.updateLines, db.listRemovedLocations);
+
+        {
+        // Search should return everyone instead of n°17
+
+        List<byte[]> indexedValuesList = Ffi.search(masterKeys.getK(), "NewLabel".getBytes(),
+        new Word[] {new Word("France")}, 0, db.fetchEntry, db.fetchChain);
+        int[] dbUids = indexedValuesBytesListToArray(indexedValuesList);
+
+        assertArrayEquals(newExpectedDbUids, dbUids);
+        }
+
+        // Check allocation problem during insertions. Allocation problem could occur when fetching entry table
+        // values
+        // whose sizes depend on words being indexed: the Entry Table Encrypted value is:
+        // `EncSym(𝐾value, (ict_uid𝑥𝑤𝑖, 𝐾𝑤𝑖 , 𝑤𝑖))`
+        for (int i = 0; i < 100; i++) {
+        Ffi.upsert(masterKeys, label, indexedValuesAndWords, db.fetchEntry, db.upsertEntry, db.upsertChain);
+        }
+
     }
 
     /*